Extract syntax type conversions into a separate module
The syntax.rs file contained a lot of code about type conversions and
environment-free expansions. To keep the file focused on the syntax
definition, this commit moves the conversion code into a separate
module, conversions.rs.
